requirements/dev.txt
-r base.txt
black==25.1.0
isort==6.0.1
ruff==0.11.0
pytest==8.3.5
pytest-django==4.10.0
factory-boy==3.3.3
ipython==9.0.2
django-debug-toolbar==4.4.6
